# significantly improve performance via parametrizing core types